Sample23

+++ Desktop/Java/Sample23.java
@@ -2,7 +2,7 @@ public class Sample23 {
   public static void main(String[] args) {
     int num1 = 5;
     int num2 = 4;
-    double div = num1/num2;
+    double div = (double)num1/(double)num2;
     System.out.println("5/4は" + div + "です。");
   }
 }
